+++ doc/CHANGES.rst
@@ -8,6 +8,8 @@ http://docs.zope.org/zope2/releases/.
 2.13.11 (unreleased)
 --------------------
 
+- Turn `UndoSupport.get_request_var_or_attr` helper into a private API.
+
 - LP #902068: Fixed missing security declaration for `ObjectManager` class.
 
 - Avoid conflicting signal registrations when run under mod_wsgi.