+++ doc/CHANGES.rst
@@ -22,6 +22,7 @@ http://docs.zope.org/zope2/releases/.
 - Updated distributions:
 
   - AccessControl = 2.13.7
+  - DocumentTemplate = 2.13.2
   - Products.BTreeFolder2 = 2.13.4
   - python-gettext = 1.2
   - repoze.who = 2.0